Starting to get a sense of things
Adding an ECS for input streams inside of ya boy Artillery. This will serve to fully indirect bristlecone and allow us to pretend cabling is the same as network input if we want to. There's a lot of advantages in logical terms here, but the big one is that now we have a really nice way to hide the playerness and localness of any entity.