Update UArtilleryAbilityMinimum.h
Add more information on intent. I can't see a way to remove the default shaping for the gameplay ability, and it may actually make the most sense to create an embodying class that actually just encapsulates the ability, wires it up, and hides everything but the stages as graphs. This honestly feels pretty elegant. Let me think about it more.