Restructure fire control machine to prep for threading
Once that's done, we can wire input down into the artillery core. i'd really really like to use enhanced input, but it doesn't have a metaphor for saving input, it's not threaded, it's very slow, and it doesn't support GameInput as its backing lib without a NASTY engine patch in 5.4 even with the plugin. Ultimately, it's just worth it for now to pipe everything through for maximum speed and control of state mutation.