Artillery frame-out interstitial
all that's left is the ability sequence layout. so the actual bit that matters.

and like 20 hours of debugging.
and jolt.
and rollback.

and then another 20 hours of debugging.

but honestly, fully threaded input registration separate from a core character with support for overlapping event mapping is pretty good.
Though I did just realize that cabling should handle detecting long and short key presses in a sane universe but... already here i guess.

good news is  cabling still has to do the keymapping. and will likely need to understand long presses too. so most of this work won't be wasted.

is that good news? I'm not sure anymore. hooboy.