Wire up but do not yet tick locomotion handlers.
This is... a lot nicer than I thought it would be as an idiom. It feels pretty good to know how, when, and why any given movement happens. The single call-site pattern was gonna bug me, but honestly, this is pretty dope. I think we'll be able to leverage this flow to really grind out some pretty impressive speed. Also, we can now easily add locomotion to something without needing an actor.

Co-Authored-By: Michael Matthews <5384199+mm64z@users.noreply.github.com>