Rework Blueprint Ability trigger to use a much more flexible input
One of the key pieces of awkwardness in GAS is that your ability gets triggered and.... Then has to rebuild all the state passed into the activate ability call or rely on access to the ability system component to regain context. I'm hoping there's not a performance reason for this, but the good news is that I'll be using POD structs.