Cut pass for time flow in artillery.
I don't _like_ the way that artillery can wait quite a while to run. I wonder if it should always run as soon as it can? But then we might skip input.

Because we do need to resim, I'm pretty okay with an artillery tick being as low as 2ms. It's not the best thing, but generally, it won't be a problem. Generally. I think. Hm.