Add Shadow Now
This allows us to freeze ticklikes in a specific moment in time, much like a frame start but with the fun ability to separate it from the current frame elsewhere _intentionally._ This is married to our ability to access the temporal shadow copies of most game sim data, and allows us to, mostly, run without any contention. wylde.