Add explicit event access for artillery abilities
I'm really perplexed why this isn't the default behavior. Honestly, huge chunks of the ability system make borderline no sense at all. Artillery will need to provide a key-check mechanism that's callable from abilities, one that can operate over an arbitrary UObject. This should be possible, given how we compute actor keys, and it honestly might be beneficial to just automate that, since almost all abilities will perform it. if we go this route, the code for this lookup should live in artillery dispatch and the calls for it should live in ActivateAbility for UArtilleryPerAbilityMinimum.