Turns out, 300 is too many.
Gotta actually get this stuff bolted down a little. This commit moves attributes out, adds a recharge phase, and defines some useful constants. It also makes sure dash doesn't fling you entirely out of the universe. Next step is creating a recharge tickable, though I'm waffling on how I want it to work. I think, unfortunately, it's going to need to be a delegate-like mechanism, and I don't know if those are threaded. The alternative is passing in a map of attribute\per-tick recharge, but then we also need some way to update that..... grasping hand is that we could have recharge values as attributes and just auto harvest them. That would mean that things Just Work which might be kinda nice. idk though.