Move update of managed transforms to game thread
I'm pretty upset about this. I had really really hoped to keep this step down in the threaded systems but I don't see a way to do so yet without a ton of surgery. This solution is pretty solid, at least as a first cut, so I'm good with it until after the demo. Ultimately, I gotta get a move on.